squelette/services/api/requirements.txt
fastapi>=0.95
uvicorn>=0.20
mlflow>=2.0
prometheus-client>=0.16
pillow>=9.0
tensorflow>=2.10
python-multipart
